+++ neo-cli/Settings.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Neo.Network.P2P;
+using Neo.SmartContract.Native;
 using System.Net;
 
 namespace Neo
@@ -33,12 +34,10 @@ namespace Neo
     internal class PathsSettings
     {
         public string Chain { get; }
-        public string Index { get; }
 
         public PathsSettings(IConfigurationSection section)
         {
-            this.Chain = string.Format(section.GetSection("Chain").Value, Message.Magic.ToString("X8"));
-            this.Index = string.Format(section.GetSection("Index").Value, Message.Magic.ToString("X8"));
+            this.Chain = string.Format(section.GetSection("Chain").Value, ProtocolSettings.Default.Magic.ToString("X8"));
         }
     }
 
@@ -48,6 +47,7 @@ namespace Neo
         public ushort WsPort { get; }
         public int MinDesiredConnections { get; }
         public int MaxConnections { get; }
+        public int MaxConnectionsPerAddress { get; }
 
         public P2PSettings(IConfigurationSection section)
         {
@@ -55,6 +55,7 @@ namespace Neo
             this.WsPort = ushort.Parse(section.GetSection("WsPort").Value);
             this.MinDesiredConnections = section.GetValue("MinDesiredConnections", Peer.DefaultMinDesiredConnections);
             this.MaxConnections = section.GetValue("MaxConnections", Peer.DefaultMaxConnections);
+            this.MaxConnectionsPerAddress = section.GetValue("MaxConnectionsPerAddress", 3);
         }
     }
 
@@ -64,6 +65,7 @@ namespace Neo
         public ushort Port { get; }
         public string SslCert { get; }
         public string SslCertPassword { get; }
+        public long MaxGasInvoke { get; }
 
         public RPCSettings(IConfigurationSection section)
         {
@@ -71,6 +73,7 @@ namespace Neo
             this.Port = ushort.Parse(section.GetSection("Port").Value);
             this.SslCert = section.GetSection("SslCert").Value;
             this.SslCertPassword = section.GetSection("SslCertPassword").Value;
+            this.MaxGasInvoke = (long)BigDecimal.Parse(section.GetValue("MaxGasInvoke", "0"), NativeContract.GAS.Decimals).Value;
         }
     }