+++ neo-cli/Shell/MainService.cs
@@ -1,15 +1,18 @@
-﻿using Akka.Actor;
+using Akka.Actor;
 using Neo.Consensus;
 using Neo.IO;
 using Neo.IO.Json;
 using Neo.Ledger;
 using Neo.Network.P2P;
+using Neo.Network.P2P.Capabilities;
 using Neo.Network.P2P.Payloads;
 using Neo.Persistence;
 using Neo.Persistence.LevelDB;
 using Neo.Plugins;
 using Neo.Services;
 using Neo.SmartContract;
+using Neo.SmartContract.Manifest;
+using Neo.SmartContract.Native;
 using Neo.VM;
 using Neo.Wallets;
 using Neo.Wallets.NEP6;
@@ -20,8 +23,8 @@ using System.IO;
 using System.IO.Compression;
 using System.Linq;
 using System.Net;
+using System.Numerics;
 using System.Security.Cryptography;
-using System.Text;
 using System.Threading.Tasks;
 using ECCurve = Neo.Cryptography.ECC.ECCurve;
 using ECPoint = Neo.Cryptography.ECC.ECPoint;
@@ -30,21 +33,11 @@ namespace Neo.Shell
 {
     internal class MainService : ConsoleServiceBase
     {
-        private const string PeerStatePath = "peers.dat";
-
         private LevelDBStore store;
         private NeoSystem system;
-        private WalletIndexer indexer;
 
         protected override string Prompt => "neo";
         public override string ServiceName => "NEO-CLI";
-
-        private WalletIndexer GetIndexer()
-        {
-            if (indexer is null)
-                indexer = new WalletIndexer(Settings.Default.Paths.Index);
-            return indexer;
-        }
 
         private static bool NoWallet()
         {
@@ -78,12 +71,10 @@ namespace Neo.Shell
                     return OnImportCommand(args);
                 case "list":
                     return OnListCommand(args);
-                case "claim":
-                    return OnClaimCommand(args);
                 case "open":
                     return OnOpenCommand(args);
-                case "rebuild":
-                    return OnRebuildCommand(args);
+                case "close":
+                    return OnCloseCommand(args);
                 case "send":
                     return OnSendCommand(args);
                 case "show":
@@ -107,41 +98,35 @@ namespace Neo.Shell
 
         private bool OnBroadcastCommand(string[] args)
         {
-            string command = args[1].ToLower();
+            if (!Enum.TryParse(args[1], true, out MessageCommand command))
+            {
+                Console.WriteLine($"Command \"{args[1]}\" is not supported.");
+                return true;
+            }
             ISerializable payload = null;
             switch (command)
             {
-                case "addr":
-                    payload = AddrPayload.Create(NetworkAddressWithTime.Create(new IPEndPoint(IPAddress.Parse(args[2]), ushort.Parse(args[3])), NetworkAddressWithTime.NODE_NETWORK, DateTime.UtcNow.ToTimestamp()));
+                case MessageCommand.Addr:
+                    payload = AddrPayload.Create(NetworkAddressWithTime.Create(IPAddress.Parse(args[2]), DateTime.UtcNow.ToTimestamp(), new FullNodeCapability(), new ServerCapability(NodeCapabilityType.TcpServer, ushort.Parse(args[3]))));
                     break;
-                case "block":
+                case MessageCommand.Block:
                     if (args[2].Length == 64 || args[2].Length == 66)
                         payload = Blockchain.Singleton.GetBlock(UInt256.Parse(args[2]));
                     else
                         payload = Blockchain.Singleton.Store.GetBlock(uint.Parse(args[2]));
                     break;
-                case "getblocks":
-                case "getheaders":
+                case MessageCommand.GetBlocks:
+                case MessageCommand.GetHeaders:
                     payload = GetBlocksPayload.Create(UInt256.Parse(args[2]));
                     break;
-                case "getdata":
-                case "inv":
+                case MessageCommand.GetData:
+                case MessageCommand.Inv:
                     payload = InvPayload.Create(Enum.Parse<InventoryType>(args[2], true), args.Skip(3).Select(UInt256.Parse).ToArray());
                     break;
-                case "tx":
+                case MessageCommand.Transaction:
                     payload = Blockchain.Singleton.GetTransaction(UInt256.Parse(args[2]));
                     break;
-                case "alert":
-                case "consensus":
-                case "filteradd":
-                case "filterload":
-                case "headers":
-                case "merkleblock":
-                case "ping":
-                case "pong":
-                case "reject":
-                case "verack":
-                case "version":
+                default:
                     Console.WriteLine($"Command \"{command}\" is not supported.");
                     return true;
             }
@@ -152,43 +137,25 @@ namespace Neo.Shell
         private bool OnDeployCommand(string[] args)
         {
             if (NoWallet()) return true;
-            var tx = LoadScriptTransaction(
+            byte[] script = LoadDeploymentScript(
                 /* filePath */ args[1],
-                /* paramTypes */ args[2],
-                /* returnType */ args[3],
-                /* hasStorage */ args[4].ToBool(),
-                /* hasDynamicInvoke */ args[5].ToBool(),
-                /* isPayable */ args[6].ToBool(),
-                /* contractName */ args[7],
-                /* contractVersion */ args[8],
-                /* contractAuthor */ args[9],
-                /* contractEmail */ args[10],
-                /* contractDescription */ args[11]);
+                /* hasStorage */ args[2].ToBool(),
+                /* isPayable */ args[3].ToBool(),
+                /* scriptHash */ out var scriptHash);
 
-            tx.Version = 1;
-            if (tx.Attributes == null) tx.Attributes = new TransactionAttribute[0];
-            if (tx.Inputs == null) tx.Inputs = new CoinReference[0];
-            if (tx.Outputs == null) tx.Outputs = new TransactionOutput[0];
-            if (tx.Witnesses == null) tx.Witnesses = new Witness[0];
-            ApplicationEngine engine = ApplicationEngine.Run(tx.Script, tx, null, true);
-            StringBuilder sb = new StringBuilder();
-            sb.AppendLine($"VM State: {engine.State}");
-            sb.AppendLine($"Gas Consumed: {engine.GasConsumed}");
-            sb.AppendLine(
-                $"Evaluation Stack: {new JArray(engine.ResultStack.Select(p => p.ToParameter().ToJson()))}");
-            Console.WriteLine(sb.ToString());
-            if (engine.State.HasFlag(VMState.FAULT))
+            Transaction tx = new Transaction { Script = script };
+            try
+            {
+                Program.Wallet.FillTransaction(tx);
+            }
+            catch (InvalidOperationException)
             {
                 Console.WriteLine("Engine faulted.");
                 return true;
             }
-
-            tx.Gas = engine.GasConsumed - Fixed8.FromDecimal(10);
-            if (tx.Gas < Fixed8.Zero) tx.Gas = Fixed8.Zero;
-            tx.Gas = tx.Gas.Ceiling();
-
-            tx = DecorateScriptTransaction(tx);
-
+            Console.WriteLine($"Script hash: {scriptHash.ToString()}");
+            Console.WriteLine($"Gas: {tx.Gas}");
+            Console.WriteLine();
             return SignAndSendTx(tx);
         }
 
@@ -207,98 +174,73 @@ namespace Neo.Shell
                 });
             }
 
-            ContractParameter[] parameters =
+            Transaction tx = new Transaction
             {
-                new ContractParameter
+                Sender = UInt160.Zero,
+                Attributes = new TransactionAttribute[0],
+                Witness = new Witness
                 {
-                    Type = ContractParameterType.String,
-                    Value = args[2]
-                },
-                new ContractParameter
-                {
-                    Type = ContractParameterType.Array,
-                    Value = contractParameters.ToArray()
+                    InvocationScript = new byte[0],
+                    VerificationScript = new byte[0]
                 }
             };
 
-            var tx = new InvocationTransaction();
-
             using (ScriptBuilder scriptBuilder = new ScriptBuilder())
             {
-                scriptBuilder.EmitAppCall(scriptHash, parameters);
-                Console.WriteLine($"Invoking script with: '{scriptBuilder.ToArray().ToHexString()}'");
+                scriptBuilder.EmitAppCall(scriptHash, args[2], contractParameters.ToArray());
                 tx.Script = scriptBuilder.ToArray();
+                Console.WriteLine($"Invoking script with: '{tx.Script.ToHexString()}'");
             }
 
-            if (tx.Attributes == null) tx.Attributes = new TransactionAttribute[0];
-            if (tx.Inputs == null) tx.Inputs = new CoinReference[0];
-            if (tx.Outputs == null) tx.Outputs = new TransactionOutput[0];
-            if (tx.Witnesses == null) tx.Witnesses = new Witness[0];
             ApplicationEngine engine = ApplicationEngine.Run(tx.Script, tx);
 
-            StringBuilder sb = new StringBuilder();
-            sb.AppendLine($"VM State: {engine.State}");
-            sb.AppendLine($"Gas Consumed: {engine.GasConsumed}");
-            sb.AppendLine(
-                $"Evaluation Stack: {new JArray(engine.ResultStack.Select(p => p.ToParameter().ToJson()))}");
-            Console.WriteLine(sb.ToString());
+            Console.WriteLine($"VM State: {engine.State}");
+            Console.WriteLine($"Gas Consumed: {engine.GasConsumed}");
+            Console.WriteLine($"Evaluation Stack: {new JArray(engine.ResultStack.Select(p => p.ToParameter().ToJson()))}");
+            Console.WriteLine();
             if (engine.State.HasFlag(VMState.FAULT))
             {
                 Console.WriteLine("Engine faulted.");
                 return true;
             }
-
-            tx = DecorateScriptTransaction(tx);
+            if (NoWallet()) return true;
+            try
+            {
+                Program.Wallet.FillTransaction(tx);
+            }
+            catch (InvalidOperationException)
+            {
+                Console.WriteLine("Error: insufficient balance.");
+                return true;
+            }
+            if (ReadUserInput("relay tx(no|yes)") != "yes")
+            {
+                return true;
+            }
             return SignAndSendTx(tx);
         }
 
-        public InvocationTransaction LoadScriptTransaction(
-            string avmFilePath, string paramTypes, string returnTypeHexString,
-            bool hasStorage, bool hasDynamicInvoke, bool isPayable,
-            string contractName, string contractVersion, string contractAuthor,
-            string contractEmail, string contractDescription)
+        private byte[] LoadDeploymentScript(string avmFilePath, bool hasStorage, bool isPayable, out UInt160 scriptHash)
         {
+            var info = new FileInfo(avmFilePath);
+            if (!info.Exists || info.Length >= Transaction.MaxTransactionSize)
+            {
+                throw new ArgumentException(nameof(avmFilePath));
+            }
+
             byte[] script = File.ReadAllBytes(avmFilePath);
-            // See ContractParameterType Enum
-            byte[] parameterList = paramTypes.HexToBytes();
-            ContractParameterType returnType = returnTypeHexString.HexToBytes()
-                .Select(p => (ContractParameterType?)p).FirstOrDefault() ?? ContractParameterType.Void;
-            ContractPropertyState properties = ContractPropertyState.NoProperty;
-            if (hasStorage) properties |= ContractPropertyState.HasStorage;
-            if (hasDynamicInvoke) properties |= ContractPropertyState.HasDynamicInvoke;
-            if (isPayable) properties |= ContractPropertyState.Payable;
+            scriptHash = script.ToScriptHash();
+            ContractFeatures properties = ContractFeatures.NoProperty;
+            if (hasStorage) properties |= ContractFeatures.HasStorage;
+            if (isPayable) properties |= ContractFeatures.Payable;
             using (ScriptBuilder sb = new ScriptBuilder())
             {
-                sb.EmitSysCall("Neo.Contract.Create", script, parameterList, returnType, properties,
-                    contractName, contractVersion, contractAuthor, contractEmail, contractDescription);
-                return new InvocationTransaction
-                {
-                    Script = sb.ToArray()
-                };
+                sb.EmitSysCall(InteropService.Neo_Contract_Create, script, properties);
+                return sb.ToArray();
             }
         }
 
-        public InvocationTransaction DecorateScriptTransaction(InvocationTransaction tx)
-        {
-            Fixed8 fee = Fixed8.FromDecimal(0.001m);
-
-            if (tx.Script.Length > 1024)
-            {
-                fee += Fixed8.FromDecimal(tx.Script.Length * 0.00001m);
-            }
-
-            return Program.Wallet.MakeTransaction(new InvocationTransaction
-            {
-                Version = tx.Version,
-                Script = tx.Script,
-                Gas = tx.Gas,
-                Attributes = tx.Attributes,
-                Inputs = tx.Inputs,
-                Outputs = tx.Outputs
-            }, fee: fee);
-        }
-
-        public bool SignAndSendTx(InvocationTransaction tx)
+        public bool SignAndSendTx(Transaction tx)
         {
             ContractParametersContext context;
             try
@@ -314,8 +256,7 @@ namespace Neo.Shell
             string msg;
             if (context.Completed)
             {
-                context.Verifiable.Witnesses = context.GetWitnesses();
-                Program.Wallet.ApplyTransaction(tx);
+                tx.Witness = context.GetWitness();
 
                 system.LocalNode.Tell(new LocalNode.Relay { Inventory = tx });
 
@@ -350,10 +291,14 @@ namespace Neo.Shell
                     Console.WriteLine("The signature is incomplete.");
                     return true;
                 }
-                context.Verifiable.Witnesses = context.GetWitnesses();
-                IInventory inventory = (IInventory)context.Verifiable;
-                system.LocalNode.Tell(new LocalNode.Relay { Inventory = inventory });
-                Console.WriteLine($"Data relay success, the hash is shown as follows:\r\n{inventory.Hash}");
+                if (!(context.Verifiable is Transaction tx))
+                {
+                    Console.WriteLine($"Only support to relay transaction.");
+                    return true;
+                }
+                tx.Witness = context.GetWitness();
+                system.LocalNode.Tell(new LocalNode.Relay { Inventory = tx });
+                Console.WriteLine($"Data relay success, the hash is shown as follows:\r\n{tx.Hash}");
             }
             catch (Exception e)
             {
@@ -474,13 +419,13 @@ namespace Neo.Shell
                 return true;
             }
             string path = args[2];
-            string password = ReadPassword("password");
+            string password = ReadUserInput("password", true);
             if (password.Length == 0)
             {
                 Console.WriteLine("cancelled");
                 return true;
             }
-            string password2 = ReadPassword("password");
+            string password2 = ReadUserInput("password", true);
             if (password != password2)
             {
                 Console.WriteLine("error");
@@ -490,7 +435,7 @@ namespace Neo.Shell
             {
                 case ".db3":
                     {
-                        Program.Wallet = UserWallet.Create(GetIndexer(), path, password);
+                        Program.Wallet = UserWallet.Create(path, password);
                         WalletAccount account = Program.Wallet.CreateAccount();
                         Console.WriteLine($"address: {account.Address}");
                         Console.WriteLine($" pubkey: {account.GetKey().PublicKey.EncodePoint(true).ToHexString()}");
@@ -500,7 +445,7 @@ namespace Neo.Shell
                     break;
                 case ".json":
                     {
-                        NEP6Wallet wallet = new NEP6Wallet(GetIndexer(), path);
+                        NEP6Wallet wallet = new NEP6Wallet(path);
                         wallet.Unlock(password);
                         WalletAccount account = wallet.CreateAccount();
                         wallet.Save();
@@ -555,7 +500,12 @@ namespace Neo.Shell
                 scriptHash = args[2].ToScriptHash();
                 path = args[3];
             }
-            string password = ReadPassword("password");
+            if (File.Exists(path))
+            {
+                Console.WriteLine($"Error: File '{path}' already exists");
+                return true;
+            }
+            string password = ReadUserInput("password", true);
             if (password.Length == 0)
             {
                 Console.WriteLine("cancelled");
@@ -590,22 +540,20 @@ namespace Neo.Shell
                 "Wallet Commands:\n" +
                 "\tcreate wallet <path>\n" +
                 "\topen wallet <path>\n" +
+                "\tclose wallet\n" +
                 "\tupgrade wallet <path>\n" +
-                "\trebuild index\n" +
                 "\tlist address\n" +
                 "\tlist asset\n" +
                 "\tlist key\n" +
-                "\tshow utxo [id|alias]\n" +
                 "\tshow gas\n" +
-                "\tclaim gas [all] [changeAddress]\n" +
                 "\tcreate address [n=1]\n" +
                 "\timport key <wif|path>\n" +
                 "\texport key [address] [path]\n" +
                 "\timport multisigaddress m pubkeys...\n" +
-                "\tsend <id|alias> <address> <value>|all [fee=0]\n" +
+                "\tsend <id|alias> <address> <value>\n" +
                 "\tsign <jsonObjectToSign>\n" +
                 "Contract Commands:\n" +
-                "\tdeploy <avmFilePath> <paramTypes> <returnTypeHexString> <hasStorage (true|false)> <hasDynamicInvoke (true|false)> <isPayable (true|false) <contractName> <contractVersion> <contractAuthor> <contractEmail> <contractDescription>\n" +
+                "\tdeploy <avmFilePath> <hasStorage (true|false)> <isPayable (true|false)\n" +
                 "\tinvoke <scripthash> <command> [optionally quoted params separated by space]\n" +
                 "Node Commands:\n" +
                 "\tshow state\n" +
@@ -652,9 +600,9 @@ namespace Neo.Shell
         {
             if (NoWallet()) return true;
 
-            if (args.Length < 5)
+            if (args.Length < 4)
             {
-                Console.WriteLine("Error. Use at least 2 public keys to create a multisig address.");
+                Console.WriteLine("Error. Invalid parameters.");
                 return true;
             }
 
@@ -696,6 +644,19 @@ namespace Neo.Shell
             catch (FormatException) { }
             if (prikey == null)
             {
+                var file = new FileInfo(args[2]);
+
+                if (!file.Exists)
+                {
+                    Console.WriteLine($"Error: File '{file.FullName}' doesn't exists");
+                    return true;
+                }
+
+                if (file.Length > 1024 * 1024)
+                {
+                    if (ReadUserInput($"The file '{file.FullName}' is too big, do you want to continue? (yes|no)", false)?.ToLowerInvariant() != "yes") return true;
+                }
+
                 string[] lines = File.ReadAllLines(args[2]);
                 for (int i = 0; i < lines.Length; i++)
                 {
@@ -737,50 +698,16 @@ namespace Neo.Shell
             }
         }
 
-        private bool OnClaimCommand(string[] args)
-        {
-            if (args.Length < 2 || args.Length > 4 || !args[1].Equals("gas", StringComparison.OrdinalIgnoreCase))
-                return base.OnCommand(args);
-
-            if (NoWallet()) return true;
-
-            bool all = args.Length > 2 && args[2].Equals("all", StringComparison.OrdinalIgnoreCase);
-            bool useChangeAddress = (all && args.Length == 4) || (!all && args.Length == 3);
-            UInt160 changeAddress = useChangeAddress ? args[args.Length - 1].ToScriptHash() : null;
-
-            if (useChangeAddress)
-            {
-                string password = ReadPassword("password");
-                if (password.Length == 0)
-                {
-                    Console.WriteLine("cancelled");
-                    return true;
-                }
-                if (!Program.Wallet.VerifyPassword(password))
-                {
-                    Console.WriteLine("Incorrect password");
-                    return true;
-                }
-            }
-
-            Coins coins = new Coins(Program.Wallet, system);
-            ClaimTransaction[] txs = all
-                ? coins.ClaimAll(changeAddress)
-                : new[] { coins.Claim(changeAddress) };
-            if (txs is null) return true;
-            foreach (ClaimTransaction tx in txs)
-                if (tx != null)
-                    Console.WriteLine($"Transaction Succeeded: {tx.Hash}");
-            return true;
-        }
-
         private bool OnShowGasCommand(string[] args)
         {
             if (NoWallet()) return true;
-
-            Coins coins = new Coins(Program.Wallet, system);
-            Console.WriteLine($"unavailable: {coins.UnavailableBonus().ToString()}");
-            Console.WriteLine($"  available: {coins.AvailableBonus().ToString()}");
+            BigInteger gas = BigInteger.Zero;
+            using (Snapshot snapshot = Blockchain.Singleton.GetSnapshot())
+                foreach (UInt160 account in Program.Wallet.GetAccounts().Select(p => p.ScriptHash))
+                {
+                    gas += NativeContract.NEO.UnclaimedGas(snapshot, account, snapshot.Height + 1);
+                }
+            Console.WriteLine($"unclaimed gas: {new BigDecimal(gas, NativeContract.GAS.Decimals)}");
             return true;
         }
 
@@ -807,19 +734,8 @@ namespace Neo.Shell
         private bool OnListAssetCommand(string[] args)
         {
             if (NoWallet()) return true;
-            foreach (var item in Program.Wallet.GetCoins().Where(p => !p.State.HasFlag(CoinState.Spent)).GroupBy(p => p.Output.AssetId, (k, g) => new
-            {
-                Asset = Blockchain.Singleton.Store.GetAssets().TryGet(k),
-                Balance = g.Sum(p => p.Output.Value),
-                Confirmed = g.Where(p => p.State.HasFlag(CoinState.Confirmed)).Sum(p => p.Output.Value)
-            }))
-            {
-                Console.WriteLine($"       id:{item.Asset.AssetId}");
-                Console.WriteLine($"     name:{item.Asset.GetName()}");
-                Console.WriteLine($"  balance:{item.Balance}");
-                Console.WriteLine($"confirmed:{item.Confirmed}");
-                Console.WriteLine();
-            }
+            Console.WriteLine($"NEO: {Program.Wallet.GetAvailable(NativeContract.NEO.Hash)}");
+            Console.WriteLine($"GAS: {Program.Wallet.GetAvailable(NativeContract.GAS.Hash)}");
             return true;
         }
 
@@ -850,7 +766,7 @@ namespace Neo.Shell
                 Console.WriteLine($"File does not exist");
                 return true;
             }
-            string password = ReadPassword("password");
+            string password = ReadUserInput("password", true);
             if (password.Length == 0)
             {
                 Console.WriteLine("cancelled");
@@ -858,7 +774,7 @@ namespace Neo.Shell
             }
             try
             {
-                Program.Wallet = OpenWallet(GetIndexer(), path, password);
+                Program.Wallet = OpenWallet(path, password);
             }
             catch (CryptographicException)
             {
@@ -869,20 +785,40 @@ namespace Neo.Shell
             return true;
         }
 
-        private bool OnRebuildCommand(string[] args)
+        /// <summary>
+        /// process "close" command
+        /// </summary>
+        /// <param name="args"></param>
+        /// <returns></returns>
+        private bool OnCloseCommand(string[] args)
         {
             switch (args[1].ToLower())
             {
-                case "index":
-                    return OnRebuildIndexCommand(args);
+                case "wallet":
+                    return OnCloseWalletCommand(args);
                 default:
                     return base.OnCommand(args);
             }
         }
 
-        private bool OnRebuildIndexCommand(string[] args)
+        /// <summary>
+        /// process "close wallet" command
+        /// </summary>
+        /// <param name="args"></param>
+        /// <returns></returns>
+        private bool OnCloseWalletCommand(string[] args)
         {
-            GetIndexer().RebuildIndex();
+            if (Program.Wallet == null)
+            {
+                Console.WriteLine($"Wallet is not opened");
+                return true;
+            }
+            Program.Wallet = null;
+            if (system.RpcServer != null)
+            {
+                system.RpcServer.Wallet = null;
+            }
+            Console.WriteLine($"Wallet is closed");
             return true;
         }
 
@@ -894,7 +830,7 @@ namespace Neo.Shell
                 return true;
             }
             if (NoWallet()) return true;
-            string password = ReadPassword("password");
+            string password = ReadUserInput("password", true);
             if (password.Length == 0)
             {
                 Console.WriteLine("cancelled");
@@ -905,72 +841,48 @@ namespace Neo.Shell
                 Console.WriteLine("Incorrect password");
                 return true;
             }
-            UIntBase assetId;
+            UInt160 assetId;
             switch (args[1].ToLower())
             {
                 case "neo":
-                case "ans":
-                    assetId = Blockchain.GoverningToken.Hash;
+                    assetId = NativeContract.NEO.Hash;
                     break;
                 case "gas":
-                case "anc":
-                    assetId = Blockchain.UtilityToken.Hash;
+                    assetId = NativeContract.GAS.Hash;
                     break;
                 default:
-                    assetId = UIntBase.Parse(args[1]);
+                    assetId = UInt160.Parse(args[1]);
                     break;
             }
-            UInt160 scriptHash = args[2].ToScriptHash();
-            bool isSendAll = string.Equals(args[3], "all", StringComparison.OrdinalIgnoreCase);
+            UInt160 to = args[2].ToScriptHash();
             Transaction tx;
-            if (isSendAll)
+            AssetDescriptor descriptor = new AssetDescriptor(assetId);
+            if (!BigDecimal.TryParse(args[3], descriptor.Decimals, out BigDecimal amount) || amount.Sign <= 0)
             {
-                Coin[] coins = Program.Wallet.FindUnspentCoins().Where(p => p.Output.AssetId.Equals(assetId)).ToArray();
-                tx = new ContractTransaction
-                {
-                    Attributes = new TransactionAttribute[0],
-                    Inputs = coins.Select(p => p.Reference).ToArray(),
-                    Outputs = new[]
-                    {
-                        new TransactionOutput
-                        {
-                            AssetId = (UInt256)assetId,
-                            Value = coins.Sum(p => p.Output.Value),
-                            ScriptHash = scriptHash
-                        }
-                    }
-                };
+                Console.WriteLine("Incorrect Amount Format");
+                return true;
             }
-            else
+            tx = Program.Wallet.MakeTransaction(null, new[]
             {
-                AssetDescriptor descriptor = new AssetDescriptor(assetId);
-                if (!BigDecimal.TryParse(args[3], descriptor.Decimals, out BigDecimal amount))
+                new TransferOutput
                 {
-                    Console.WriteLine("Incorrect Amount Format");
-                    return true;
+                    AssetId = assetId,
+                    Value = amount,
+                    ScriptHash = to
                 }
-                Fixed8 fee = args.Length >= 5 ? Fixed8.Parse(args[4]) : Fixed8.Zero;
-                tx = Program.Wallet.MakeTransaction(null, new[]
-                {
-                    new TransferOutput
-                    {
-                        AssetId = assetId,
-                        Value = amount,
-                        ScriptHash = scriptHash
-                    }
-                }, fee: fee);
-                if (tx == null)
-                {
-                    Console.WriteLine("Insufficient funds");
-                    return true;
-                }
+            });
+
+            if (tx == null)
+            {
+                Console.WriteLine("Insufficient funds");
+                return true;
             }
+
             ContractParametersContext context = new ContractParametersContext(tx);
             Program.Wallet.Sign(context);
             if (context.Completed)
             {
-                tx.Witnesses = context.GetWitnesses();
-                Program.Wallet.ApplyTransaction(tx);
+                tx.Witness = context.GetWitness();
                 system.LocalNode.Tell(new LocalNode.Relay { Inventory = tx });
                 Console.WriteLine($"TXID: {tx.Hash}");
             }
@@ -979,6 +891,7 @@ namespace Neo.Shell
                 Console.WriteLine("SignatureContext:");
                 Console.WriteLine(context.ToString());
             }
+
             return true;
         }
 
@@ -992,8 +905,6 @@ namespace Neo.Shell
                     return OnShowPoolCommand(args);
                 case "state":
                     return OnShowStateCommand(args);
-                case "utxo":
-                    return OnShowUtxoCommand(args);
                 default:
                     return base.OnCommand(args);
             }
@@ -1009,10 +920,10 @@ namespace Neo.Shell
                     out IEnumerable<Transaction> unverifiedTransactions);
                 Console.WriteLine("Verified Transactions:");
                 foreach (Transaction tx in verifiedTransactions)
-                    Console.WriteLine($" {tx.Hash} {tx.GetType().Name}");
+                    Console.WriteLine($" {tx.Hash} {tx.GetType().Name} {tx.NetworkFee} GAS_NetFee {tx.IsLowPriority}");
                 Console.WriteLine("Unverified Transactions:");
                 foreach (Transaction tx in unverifiedTransactions)
-                    Console.WriteLine($" {tx.Hash} {tx.GetType().Name}");
+                    Console.WriteLine($" {tx.Hash} {tx.GetType().Name} {tx.NetworkFee} GAS_NetFee {tx.IsLowPriority}");
             }
             Console.WriteLine($"total: {Blockchain.Singleton.MemPool.Count}, verified: {Blockchain.Singleton.MemPool.VerifiedCount}, unverified: {Blockchain.Singleton.MemPool.UnVerifiedCount}");
             return true;
@@ -1028,16 +939,12 @@ namespace Neo.Shell
                 while (!stop)
                 {
                     Console.SetCursorPosition(0, 0);
-                    uint wh = 0;
-                    if (Program.Wallet != null)
-                        wh = (Program.Wallet.WalletHeight > 0) ? Program.Wallet.WalletHeight - 1 : 0;
-
-                    WriteLineWithoutFlicker($"block: {wh}/{Blockchain.Singleton.Height}/{Blockchain.Singleton.HeaderHeight}  connected: {LocalNode.Singleton.ConnectedCount}  unconnected: {LocalNode.Singleton.UnconnectedCount}");
+                    WriteLineWithoutFlicker($"block: {Blockchain.Singleton.Height}/{Blockchain.Singleton.HeaderHeight}  connected: {LocalNode.Singleton.ConnectedCount}  unconnected: {LocalNode.Singleton.UnconnectedCount}");
                     int linesWritten = 1;
                     foreach (RemoteNode node in LocalNode.Singleton.GetRemoteNodes().Take(Console.WindowHeight - 2).ToArray())
                     {
                         WriteLineWithoutFlicker(
-                            $"  ip: {node.Remote.Address}\tport: {node.Remote.Port}\tlisten: {node.ListenerPort}\theight: {node.Version?.StartHeight}");
+                            $"  ip: {node.Remote.Address}\tport: {node.Remote.Port}\tlisten: {node.ListenerTcpPort}\theight: {node.LastBlockIndex}");
                         linesWritten++;
                     }
 
@@ -1051,39 +958,6 @@ namespace Neo.Shell
             task.Wait();
             Console.WriteLine();
             Console.CursorVisible = true;
-            return true;
-        }
-
-        private bool OnShowUtxoCommand(string[] args)
-        {
-            if (NoWallet()) return true;
-            IEnumerable<Coin> coins = Program.Wallet.FindUnspentCoins();
-            if (args.Length >= 3)
-            {
-                UInt256 assetId;
-                switch (args[2].ToLower())
-                {
-                    case "neo":
-                    case "ans":
-                        assetId = Blockchain.GoverningToken.Hash;
-                        break;
-                    case "gas":
-                    case "anc":
-                        assetId = Blockchain.UtilityToken.Hash;
-                        break;
-                    default:
-                        assetId = UInt256.Parse(args[2]);
-                        break;
-                }
-                coins = coins.Where(p => p.Output.AssetId.Equals(assetId));
-            }
-            Coin[] coins_array = coins.ToArray();
-            const int MAX_SHOW = 100;
-            for (int i = 0; i < coins_array.Length && i < MAX_SHOW; i++)
-                Console.WriteLine($"{coins_array[i].Reference.PrevHash}:{coins_array[i].Reference.PrevIndex}");
-            if (coins_array.Length > MAX_SHOW)
-                Console.WriteLine($"({coins_array.Length - MAX_SHOW} more)");
-            Console.WriteLine($"total: {coins_array.Length} UTXOs");
             return true;
         }
 
@@ -1101,12 +975,19 @@ namespace Neo.Shell
                 }
             store = new LevelDBStore(Path.GetFullPath(Settings.Default.Paths.Chain));
             system = new NeoSystem(store);
-            system.StartNode(Settings.Default.P2P.Port, Settings.Default.P2P.WsPort, Settings.Default.P2P.MinDesiredConnections, Settings.Default.P2P.MaxConnections);
+            system.StartNode(new ChannelsConfig
+            {
+                Tcp = new IPEndPoint(IPAddress.Any, Settings.Default.P2P.Port),
+                WebSocket = new IPEndPoint(IPAddress.Any, Settings.Default.P2P.WsPort),
+                MinDesiredConnections = Settings.Default.P2P.MinDesiredConnections,
+                MaxConnections = Settings.Default.P2P.MaxConnections,
+                MaxConnectionsPerAddress = Settings.Default.P2P.MaxConnectionsPerAddress
+            });
             if (Settings.Default.UnlockWallet.IsActive)
             {
                 try
                 {
-                    Program.Wallet = OpenWallet(GetIndexer(), Settings.Default.UnlockWallet.Path, Settings.Default.UnlockWallet.Password);
+                    Program.Wallet = OpenWallet(Settings.Default.UnlockWallet.Path, Settings.Default.UnlockWallet.Password);
                 }
                 catch (CryptographicException)
                 {
@@ -1123,7 +1004,8 @@ namespace Neo.Shell
                     Settings.Default.RPC.Port,
                     wallet: Program.Wallet,
                     sslCert: Settings.Default.RPC.SslCert,
-                    password: Settings.Default.RPC.SslCertPassword);
+                    password: Settings.Default.RPC.SslCertPassword,
+                    maxGasInvoke: Settings.Default.RPC.MaxGasInvoke);
             }
         }
 
@@ -1170,15 +1052,35 @@ namespace Neo.Shell
                 Console.WriteLine("error");
                 return true;
             }
+
+            bool isTemp;
+            string fileName;
             var pluginName = args[1];
-            var address = string.Format(Settings.Default.PluginURL, pluginName, typeof(Plugin).Assembly.GetVersion());
-            var fileName = Path.Combine("Plugins", $"{pluginName}.zip");
-            Directory.CreateDirectory("Plugins");
-            Console.WriteLine($"Downloading from {address}");
-            using (WebClient wc = new WebClient())
+
+            if (!File.Exists(pluginName))
             {
-                wc.DownloadFile(address, fileName);
+                if (string.IsNullOrEmpty(Settings.Default.PluginURL))
+                {
+                    Console.WriteLine("You must define `PluginURL` in your `config.json`");
+                    return true;
+                }
+
+                var address = string.Format(Settings.Default.PluginURL, pluginName, typeof(Plugin).Assembly.GetVersion());
+                fileName = Path.Combine(Path.GetTempPath(), $"{pluginName}.zip");
+                isTemp = true;
+
+                Console.WriteLine($"Downloading from {address}");
+                using (WebClient wc = new WebClient())
+                {
+                    wc.DownloadFile(address, fileName);
+                }
             }
+            else
+            {
+                fileName = pluginName;
+                isTemp = false;
+            }
+
             try
             {
                 ZipFile.ExtractToDirectory(fileName, ".");
@@ -1190,8 +1092,12 @@ namespace Neo.Shell
             }
             finally
             {
-                File.Delete(fileName);
+                if (isTemp)
+                {
+                    File.Delete(fileName);
+                }
             }
+
             Console.WriteLine($"Install successful, please restart neo-cli.");
             return true;
         }
@@ -1203,8 +1109,20 @@ namespace Neo.Shell
                 Console.WriteLine("error");
                 return true;
             }
+
             var pluginName = args[1];
-            Directory.Delete(Path.Combine("Plugins", pluginName), true);
+
+            if (!Plugin.Plugins.Any(u => u.Name == pluginName))
+            {
+                Console.WriteLine("Plugin not found");
+                return true;
+            }
+
+            if (Directory.Exists(Path.Combine("Plugins", pluginName)))
+            {
+                Directory.Delete(Path.Combine("Plugins", pluginName), true);
+            }
+
             File.Delete(Path.Combine("Plugins", $"{pluginName}.dll"));
             Console.WriteLine($"Uninstall successful, please restart neo-cli.");
             return true;
@@ -1228,27 +1146,32 @@ namespace Neo.Shell
                 Console.WriteLine("File does not exist.");
                 return true;
             }
-            string password = ReadPassword("password");
+            string password = ReadUserInput("password", true);
             if (password.Length == 0)
             {
                 Console.WriteLine("cancelled");
                 return true;
             }
+            if (File.Exists(path_new))
+            {
+                Console.WriteLine($"File '{path_new}' already exists");
+                return true;
+            }
             string path_new = Path.ChangeExtension(path, ".json");
-            NEP6Wallet.Migrate(GetIndexer(), path_new, path, password).Save();
+            NEP6Wallet.Migrate(path_new, path, password).Save();
             Console.WriteLine($"Wallet file upgrade complete. New wallet file has been auto-saved at: {path_new}");
             return true;
         }
 
-        private static Wallet OpenWallet(WalletIndexer indexer, string path, string password)
+        private static Wallet OpenWallet(string path, string password)
         {
             if (Path.GetExtension(path) == ".db3")
             {
-                return UserWallet.Open(indexer, path, password);
+                return UserWallet.Open(path, password);
             }
             else
             {
-                NEP6Wallet nep6wallet = new NEP6Wallet(indexer, path);
+                NEP6Wallet nep6wallet = new NEP6Wallet(path);
                 nep6wallet.Unlock(password);
                 return nep6wallet;
             }